+++ packages/react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ponder/react
 
+## 0.9.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.9.24
+
+## 0.9.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.9.23
+
+## 0.9.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.9.22
+
 ## 0.9.21
 
 ### Patch Changes